+++ revenuecat_examples/MagicWeather/lib/main_amazon_appstore.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:magic_weather_flutter/src/app.dart';
 
-import 'flavor_config.dart';
+import 'store_config.dart';
 
 void main() {
-  FlavorConfig(
-    flavor: Flavor.amazonAppstore,
+  StoreConfig(
+    store: Store.amazonAppstore,
     //TO DO: add the API key for your app from the RevenueCat dashboard: https://app.revenuecat.com
     apiKey: "",
   );
