+++ revenuecat_examples/MagicWeather/lib/src/views/home.dart
@@ -7,7 +7,7 @@ import 'package:purchases_flutter/purchases_flutter.dart';
 import 'package:magic_weather_flutter/src/model/singletons_data.dart';
 import 'package:magic_weather_flutter/src/model/styles.dart';
 
-import '../../flavor_config.dart';
+import '../../store_config.dart';
 
 final GlobalKey<NavigatorState> firstTabNavKey = GlobalKey<NavigatorState>();
 final GlobalKey<NavigatorState> secondTabNavKey = GlobalKey<NavigatorState>();
@@ -38,12 +38,12 @@ class AppContainerState extends State<AppContainer> {
     - observerMode is false, so Purchases will automatically handle finishing transactions. Read more about Observer Mode here: https://docs.revenuecat.com/docs/observer-mode
     */
     PurchasesConfiguration configuration;
-    if (FlavorConfig.isForAmazonAppstore()) {
-      configuration = AmazonConfiguration(FlavorConfig.instance.apiKey)
+    if (StoreConfig.isForAmazonAppstore()) {
+      configuration = AmazonConfiguration(StoreConfig.instance.apiKey)
         ..appUserID = null
         ..observerMode = false;
     } else {
-      configuration = PurchasesConfiguration(FlavorConfig.instance.apiKey)
+      configuration = PurchasesConfiguration(StoreConfig.instance.apiKey)
         ..appUserID = null
         ..observerMode = false;
     }
