+++ revenuecat_examples/MagicWeather/lib/store_config.dart
