+++ revenuecat_examples/purchase_tester/lib/main_amazon_appstore.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:purchases_flutter_example/src/constant.dart';
 
-import 'flavor_config.dart';
+import 'store_config.dart';
 import 'src/app.dart';
 
 void main() {
